bookbrain-backend/pyproject.toml
[project]
name = "bookbrain-backend"
version = "0.1.0"
description = "BookBrain API - Personal Knowledge Search System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "psycopg-pool>=3.3.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "qdrant-client>=1.16.2",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.14.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
markers = [
    "integration: marks tests as integration tests (require Docker)",
]

[tool.ruff]
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bookbrain"]
